put filter-button in a separate class
