+++ main.js
@@ -1,31 +1,16 @@
 import { TopThreeView } from './challenges.js';
 import { ChallengeListView } from './challenges.js';
-import { openPopup } from './mobile-menu.js';
 import { FilterUI } from './filterUI.js';
-import { FilterLogic } from './filter-logic.js';
+import { FilterButton } from './filter-button.js';
+import { MobileMenu } from './mobile-menu.js';
 
 const isOnMainPage = document.querySelector('.main-page');
 const isOnChallengeSite = document.querySelector('.challenges-site');
 
-// HTML elements
-const filterDiv = document.querySelector('.filterDiv');
-const filterButton = document.querySelector('.filterButton');
-const closeMenu = document.querySelector('.closeMenu');
 
-// Event listeners
-document.querySelector('.navbar-button').addEventListener('click', openPopup);
 
-// Filter button and close filter
-if (!isOnMainPage) {
-  filterButton.addEventListener('click', () => {
-    filterDiv.style.display = 'block';
-    filterButton.style.display = 'none';
-  });
-  closeMenu.addEventListener('click', () => {
-    filterDiv.style.display = 'none';
-    filterButton.style.display = 'block';
-  });
-}
+
+
 
 // ███████████████ Entry Point ███████████████ -------------------------------------------------------------------------
 
@@ -36,6 +21,7 @@ if (isOnChallengeSite) {
   let view = new ChallengeListView();
   view.render(challengesContainer);
   new FilterUI().render(challengesContainer);
+  new FilterButton().render();
 }
 
 if (isOnMainPage) {
@@ -43,4 +29,5 @@ if (isOnMainPage) {
     '.challenges-container.main-page'
   );
   new TopThreeView().render(topThreeContainer);
+  const mobileMenu = new MobileMenu
 }
